pyproject.toml
[project]
name = "repl-nix-llamaindex"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "llama-index==0.10.0",
    "python-dotenv>=1.0.1",
    "langchain>=0.0.303",
    "openai>=1.0.0",
    "gpt-index>=0.8.42",
    "llama-index-llms-openai>=0.1.31",
    "llama-index-core>=0.10.68.post1",
]
